Refactored detection of files a lot and added possibility to have a show but not automatically trash watched episodes. Started a bit on adding calendar and browsing filesystem apps

+++ web_frontend/CMi/tvshows/sort_eps.py
@@ -23,8 +23,9 @@ date_regexs = [
 
 def delete_watched_episodes():
     for episode in Episode.objects.filter(watched=True):
-        send_to_trash(episode.filepath)
-        episode.delete()
+        if episode.show.auto_erase:
+            send_to_trash(episode.filepath)
+            episode.delete()
     
 def clean_episode_db():
     #print 'cleaning episode db...'
